Add initialization options to the Path class
This introduces different options for the instantiation of the path
class. `Path.Polynomial` implements the previous interface to generate
a polynomial interpolator through a list of positions. The new
`Path.from_points` instead uses splines to create an interpolator
through a list of points. In all cases, the longitudinal position `s`
is now approximated by a monotonic spline instead of a polynomial.

numpy is no longer needed as a dependency to geothermsim.

interpax is now needed as a dependency.

Functions that encapsulate the creation of a borefield and the
calculation of the *g*-function are now jittable and compatible with
automatic differentiation.